refactor(pkg/trie): Move StorageHandle to node_storage and improve godoc
